requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.1
cytoolz==0.12.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.4
install==1.3.5
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
names==0.3.0
netaddr==0.8.0
numpy==1.23.4
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.1.221024
parsimonious==0.8.1
Pillow==9.3.0
protobuf==3.19.5
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pywin32==304
requests==2.28.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
yarl==1.8.1
